completed all exercise

+++ part4/bloglist/controllers/blogs.js
@@ -1,8 +1,12 @@
 const blogRouter = require('express').Router()
-const Blog = require('../models/blog.js')
+const Blog = require('../models/blog')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
+// const middleware = require('../utils/middleware')
+
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -15,24 +19,51 @@ blogRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogRouter.post('/', async (request, response) => {
+
+blogRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
+
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user.id
   })
 
+
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 
 })
 
-blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogRouter.delete('/:id', userExtractor, async (request, response) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing' })
+  }
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  // const user = await User.findById(decodedToken.id)
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+
+  if ( blog.user._id.toString() === user._id.toString() ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+    console.log('content removed')
+  }
+
 })
 
 blogRouter.put('/:id', async (request, response) => {
